Moved useeffect to its own hook
